backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# AI and ML libraries
openai==1.12.0
google-generativeai==0.3.2
chromadb==0.4.18

# Document processing
PyPDF2==3.0.1
PyMuPDF==1.23.8
python-docx==1.1.0

# HTTP client
httpx==0.27.0
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# CORS middleware
python-cors==1.7.0

# Logging
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Optional: Redis for caching


# Optional: Monitoring
prometheus-client==0.19.0
